reorganize transformation tests
