+++ astropy/coordinates/tests/test_intermediate_transformations.py
@@ -1,0 +1,321 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""Accuracy tests for GCRS coordinate transformations, primarily to/from AltAz.
+
+"""
+from __future__ import (absolute_import, division, print_function,
+                        unicode_literals)
+
+import numpy as np
+
+from ... import units as u
+from ...tests.helper import (pytest,
+                             quantity_allclose as allclose,
+                             assert_quantity_allclose as assert_allclose)
+from ...time import Time
+from .. import (EarthLocation, get_sun, ICRS, GCRS, CIRS, ITRS, AltAz,
+                PrecessedGeocentric, CartesianRepresentation,
+                SphericalRepresentation)
+
+from .utils import randomly_sample_sphere
+
+def test_icrs_cirs():
+    """
+    Check a few cases of ICRS<->CIRS for consistency.
+
+    Also includes the CIRS<->CIRS transforms at different times, as those go
+    through ICRS
+    """
+    ra, dec, dist = randomly_sample_sphere(200)
+    inod = ICRS(ra=ra, dec=dec)
+    iwd = ICRS(ra=ra, dec=dec, distance=dist*u.pc)
+
+    cframe1 = CIRS()
+    cirsnod = inod.transform_to(cframe1)  #uses the default time
+    #first do a round-tripping test
+    inod2 = cirsnod.transform_to(ICRS)
+    assert_allclose(inod.ra, inod2.ra)
+    assert_allclose(inod.dec, inod2.dec)
+
+    #now check that a different time yields different answers
+    cframe2 = CIRS(obstime=Time('J2005', scale='utc'))
+    cirsnod2 = inod.transform_to(cframe2)
+    assert not allclose(cirsnod.ra, cirsnod2.ra, rtol=1e-8)
+    assert not allclose(cirsnod.dec, cirsnod2.dec, rtol=1e-8)
+
+    # parallax effects should be included, so with and w/o distance should be different
+    cirswd = iwd.transform_to(cframe1)
+    assert not allclose(cirswd.ra, cirsnod.ra, rtol=1e-8)
+    assert not allclose(cirswd.dec, cirsnod.dec, rtol=1e-8)
+    # and the distance should transform at least somehow
+    assert not allclose(cirswd.distance, iwd.distance, rtol=1e-8)
+
+    #now check that the cirs self-transform works as expected
+    cirsnod3 = cirsnod.transform_to(cframe1)  # should be a no-op
+    assert_allclose(cirsnod.ra, cirsnod3.ra)
+    assert_allclose(cirsnod.dec, cirsnod3.dec)
+
+    cirsnod4 = cirsnod.transform_to(cframe2)  # should be different
+    assert not allclose(cirsnod4.ra, cirsnod.ra, rtol=1e-8)
+    assert not allclose(cirsnod4.dec, cirsnod.dec, rtol=1e-8)
+
+    cirsnod5 = cirsnod4.transform_to(cframe1)  # should be back to the same
+    assert_allclose(cirsnod.ra, cirsnod5.ra)
+    assert_allclose(cirsnod.dec, cirsnod5.dec)
+
+
+def test_icrs_gcrs():
+    """
+    Check ICRS<->GCRS for consistency
+    """
+    ra, dec, dist = randomly_sample_sphere(200)
+    inod = ICRS(ra=ra, dec=dec)
+    iwd = ICRS(ra=ra, dec=dec, distance=dist*u.pc)
+
+    gframe1 = GCRS()
+    gcrsnod = inod.transform_to(gframe1)  #uses the default time
+    #first do a round-tripping test
+    inod2 = gcrsnod.transform_to(ICRS)
+    assert_allclose(inod.ra, inod2.ra)
+    assert_allclose(inod.dec, inod2.dec)
+
+    #now check that a different time yields different answers
+    gframe2 = GCRS(obstime=Time('J2005', scale='utc'))
+    gcrsnod2 = inod.transform_to(gframe2)
+    assert not allclose(gcrsnod.ra, gcrsnod2.ra, rtol=1e-8, atol=1e-10*u.deg)
+    assert not allclose(gcrsnod.dec, gcrsnod2.dec, rtol=1e-8, atol=1e-10*u.deg)
+
+    # parallax effects should be included, so with and w/o distance should be different
+    gcrswd = iwd.transform_to(gframe1)
+    assert not allclose(gcrswd.ra, gcrsnod.ra, rtol=1e-8, atol=1e-10*u.deg)
+    assert not allclose(gcrswd.dec, gcrsnod.dec, rtol=1e-8, atol=1e-10*u.deg)
+    # and the distance should transform at least somehow
+    assert not allclose(gcrswd.distance, iwd.distance, rtol=1e-8,
+                        atol=1e-10*u.pc)
+
+    #now check that the cirs self-transform works as expected
+    gcrsnod3 = gcrsnod.transform_to(gframe1)  # should be a no-op
+    assert_allclose(gcrsnod.ra, gcrsnod3.ra)
+    assert_allclose(gcrsnod.dec, gcrsnod3.dec)
+
+    gcrsnod4 = gcrsnod.transform_to(gframe2)  # should be different
+    assert not allclose(gcrsnod4.ra, gcrsnod.ra, rtol=1e-8, atol=1e-10*u.deg)
+    assert not allclose(gcrsnod4.dec, gcrsnod.dec, rtol=1e-8, atol=1e-10*u.deg)
+
+    gcrsnod5 = gcrsnod4.transform_to(gframe1)  # should be back to the same
+    assert_allclose(gcrsnod.ra, gcrsnod5.ra, rtol=1e-8, atol=1e-10*u.deg)
+    assert_allclose(gcrsnod.dec, gcrsnod5.dec, rtol=1e-8, atol=1e-10*u.deg)
+
+    #also make sure that a GCRS with a different geoloc/geovel gets a different answer
+    # roughly a moon-like frame
+    gframe3 = GCRS(obsgeoloc=[385000., 0, 0]*u.km, obsgeovel=[1, 0, 0]*u.km/u.s)
+    gcrsnod6 = inod.transform_to(gframe3)  # should be different
+    assert not allclose(gcrsnod.ra, gcrsnod6.ra, rtol=1e-8, atol=1e-10*u.deg)
+    assert not allclose(gcrsnod.dec, gcrsnod6.dec, rtol=1e-8, atol=1e-10*u.deg)
+    inodviag3 = gcrsnod6.transform_to(ICRS)  # and now back to the original
+    assert_allclose(inod.ra, inodviag3.ra)
+    assert_allclose(inod.dec, inodviag3.dec)
+
+
+def test_cirs_to_altaz():
+    """
+    Check the basic CIRS<->AltAz transforms.  More thorough checks implicitly
+    happen in `test_iau_fullstack`
+    """
+    from .. import EarthLocation
+
+    ra, dec, dist = randomly_sample_sphere(200)
+    cirs = CIRS(ra=ra, dec=dec, obstime='J2000')
+    crepr = SphericalRepresentation(lon=ra, lat=dec, distance=dist)
+    cirscart = CIRS(crepr, obstime=cirs.obstime, representation=CartesianRepresentation)
+
+    loc = EarthLocation(lat=0*u.deg, lon=0*u.deg, height=0*u.m)
+    altazframe = AltAz(location=loc, obstime=Time('J2005'))
+
+    cirs2 = cirs.transform_to(altazframe).transform_to(cirs)
+    cirs3 = cirscart.transform_to(altazframe).transform_to(cirs)
+
+    #check round-tripping
+    assert_allclose(cirs.ra, cirs2.ra)
+    assert_allclose(cirs.dec, cirs2.dec)
+    assert_allclose(cirs.ra, cirs3.ra)
+    assert_allclose(cirs.dec, cirs3.dec)
+
+
+def test_gcrs_itrs():
+    """
+    Check basic GCRS<->ITRS transforms for round-tripping.
+    """
+    ra, dec, _ = randomly_sample_sphere(200)
+    gcrs = GCRS(ra=ra, dec=dec, obstime='J2000')
+    gcrs6 = GCRS(ra=ra, dec=dec, obstime='J2006')
+
+    gcrs2 = gcrs.transform_to(ITRS).transform_to(gcrs)
+    gcrs6_2 = gcrs6.transform_to(ITRS).transform_to(gcrs)
+
+    assert_allclose(gcrs.ra, gcrs2.ra)
+    assert_allclose(gcrs.dec, gcrs2.dec)
+    assert not allclose(gcrs.ra, gcrs6_2.ra)
+    assert not allclose(gcrs.dec, gcrs6_2.dec)
+
+    #also try with the cartesian representation
+    gcrsc = gcrs.realize_frame(gcrs.data)
+    gcrsc.representation = CartesianRepresentation
+    gcrsc2 = gcrsc.transform_to(ITRS).transform_to(gcrsc)
+    assert_allclose(gcrsc.spherical.lon.deg, gcrsc2.ra.deg)
+    assert_allclose(gcrsc.spherical.lat, gcrsc2.dec)
+
+
+def test_cirs_itrs():
+    """
+    Check basic CIRS<->ITRS transforms for round-tripping.
+    """
+    ra, dec, _ = randomly_sample_sphere(200)
+    cirs = CIRS(ra=ra, dec=dec, obstime='J2000')
+    cirs6 = CIRS(ra=ra, dec=dec, obstime='J2006')
+
+    cirs2 = cirs.transform_to(ITRS).transform_to(cirs)
+    cirs6_2 = cirs6.transform_to(ITRS).transform_to(cirs) # different obstime
+
+    #just check round-tripping
+    assert_allclose(cirs.ra, cirs2.ra)
+    assert_allclose(cirs.dec, cirs2.dec)
+    assert not allclose(cirs.ra, cirs6_2.ra)
+    assert not allclose(cirs.dec, cirs6_2.dec)
+
+
+def test_gcrs_cirs():
+    """
+    Check GCRS<->CIRS transforms for round-tripping.  More complicated than the
+    above two because it's multi-hop
+    """
+    ra, dec, _ = randomly_sample_sphere(200)
+    gcrs = GCRS(ra=ra, dec=dec, obstime='J2000')
+    gcrs6 = GCRS(ra=ra, dec=dec, obstime='J2006')
+
+    gcrs2 = gcrs.transform_to(CIRS).transform_to(gcrs)
+    gcrs6_2 = gcrs6.transform_to(CIRS).transform_to(gcrs)
+
+    assert_allclose(gcrs.ra, gcrs2.ra)
+    assert_allclose(gcrs.dec, gcrs2.dec)
+    assert not allclose(gcrs.ra, gcrs6_2.ra)
+    assert not allclose(gcrs.dec, gcrs6_2.dec)
+
+    #now try explicit intermediate pathways and ensure they're all consistent
+    gcrs3 = gcrs.transform_to(ITRS).transform_to(CIRS).transform_to(ITRS).transform_to(gcrs)
+    assert_allclose(gcrs.ra, gcrs3.ra)
+    assert_allclose(gcrs.dec, gcrs3.dec)
+
+    gcrs4 = gcrs.transform_to(ICRS).transform_to(CIRS).transform_to(ICRS).transform_to(gcrs)
+    assert_allclose(gcrs.ra, gcrs4.ra)
+    assert_allclose(gcrs.dec, gcrs4.dec)
+
+
+def test_gcrs_altaz():
+    """
+    Check GCRS<->AltAz transforms for round-tripping.  Has multiple paths
+    """
+    from .. import EarthLocation
+
+    ra, dec, _ = randomly_sample_sphere(1)
+    gcrs = GCRS(ra=ra[0], dec=dec[0], obstime='J2000')
+
+    # check array times sure N-d arrays work
+    times = Time(np.linspace(2456293.25, 2456657.25, 51) * u.day,
+                 format='jd', scale='utc')
+
+    loc = EarthLocation(lon=10 * u.deg, lat=80. * u.deg)
+    aaframe = AltAz(obstime=times, location=loc)
+
+    aa1 = gcrs.transform_to(aaframe)
+    aa2 = gcrs.transform_to(ICRS).transform_to(CIRS).transform_to(aaframe)
+    aa3 = gcrs.transform_to(ITRS).transform_to(CIRS).transform_to(aaframe)
+
+    # make sure they're all consistent
+    assert_allclose(aa1.alt, aa2.alt)
+    assert_allclose(aa1.az, aa2.az)
+    assert_allclose(aa1.alt, aa3.alt)
+    assert_allclose(aa1.az, aa3.az)
+
+
+def test_precessed_geocentric():
+    assert PrecessedGeocentric().equinox.jd == Time('J2000', scale='utc').jd
+
+    gcrs_coo = GCRS(180*u.deg, 2*u.deg, distance=10000*u.km)
+    pgeo_coo = gcrs_coo.transform_to(PrecessedGeocentric)
+    assert np.abs(gcrs_coo.ra - pgeo_coo.ra) > 10*u.marcsec
+    assert np.abs(gcrs_coo.dec - pgeo_coo.dec) > 10*u.marcsec
+    assert_allclose(gcrs_coo.distance, pgeo_coo.distance)
+
+    gcrs_roundtrip = pgeo_coo.transform_to(GCRS)
+    assert_allclose(gcrs_coo.ra, gcrs_roundtrip.ra)
+    assert_allclose(gcrs_coo.dec, gcrs_roundtrip.dec)
+    assert_allclose(gcrs_coo.distance, gcrs_roundtrip.distance)
+
+    pgeo_coo2 = gcrs_coo.transform_to(PrecessedGeocentric(equinox='B1850'))
+    assert np.abs(gcrs_coo.ra - pgeo_coo2.ra) > 1.5*u.deg
+    assert np.abs(gcrs_coo.dec - pgeo_coo2.dec) > 0.5*u.deg
+    assert_allclose(gcrs_coo.distance, pgeo_coo2.distance)
+
+    gcrs2_roundtrip = pgeo_coo2.transform_to(GCRS)
+    assert_allclose(gcrs_coo.ra, gcrs2_roundtrip.ra)
+    assert_allclose(gcrs_coo.dec, gcrs2_roundtrip.dec)
+    assert_allclose(gcrs_coo.distance, gcrs2_roundtrip.distance)
+
+
+# shared by parametrized tests below
+std_altazs = [AltAz(location=EarthLocation(-90*u.deg, 65*u.deg),
+                    obstime=Time('J2000')), #J2000 is often a default so this might work when others don't
+              AltAz(location=EarthLocation(-90*u.deg, 65*u.deg),
+                    obstime=Time('2014-01-01 00:00:00')),
+              AltAz(location=EarthLocation(-90*u.deg, 65*u.deg),
+                    obstime=Time('2014-08-01 08:00:00')),
+              AltAz(location=EarthLocation(120*u.deg, -35*u.deg),
+                    obstime=Time('2014-01-01 00:00:00'))
+              ]
+
+@pytest.mark.parametrize('altaz', std_altazs)
+def test_gcrs_altaz_sun(altaz):
+    """
+    Sanity-check that the sun is at a reasonable distance from any altaz
+    """
+    earthecc = 0.017  # roughly earth orbital eccentricity
+
+    sun = get_sun(altaz.obstime)
+    assert sun.frame.name == 'gcrs'
+
+    # the .to(u.au) is not necessary, it just makes the asserts on failure more readable
+    assert sun.distance.to(u.au) < (earthecc + 1)*u.au
+    assert sun.distance.to(u.au) > (earthecc - 1)*u.au
+
+    sunaa = sun.transform_to(altaz)
+    assert sunaa.distance.to(u.au) < (earthecc + 1)*u.au
+    assert sunaa.distance.to(u.au) > (earthecc - 1)*u.au
+
+
+@pytest.mark.parametrize('altaz', std_altazs)
+def test_gcrs_altaz_moonish(altaz):
+    """
+    Sanity-check that an object resembling the moon goes to the right place
+    """
+    moondist = 385000*u.km  # approximate moon semi-major orbit axis of moon
+    cart = CartesianRepresentation(3**-0.5*moondist, 3**-0.5*moondist, 3**-0.5*moondist)
+    moon = GCRS(cart, obstime=altaz.obstime)
+
+    moonaa = moon.transform_to(altaz)
+    assert np.abs(moonaa.distance - moon.distance).to(u.km) > 1000*u.km
+    assert np.abs(moonaa.distance - moon.distance).to(u.km) < 7000*u.km
+
+    #also should add checks that the alt/az are different for different earth locations
+
+@pytest.mark.parametrize('altaz', std_altazs)
+def test_cirs_altaz_moonish(altaz):
+    """
+    Sanity-check that an object resembling the moon goes to the right place if starting from CIRS
+    """
+    moondist = 385000*u.km  # approximate moon semi-major orbit axis of moon
+    cart = CartesianRepresentation(3**-0.5*moondist, 3**-0.5*moondist, 3**-0.5*moondist)
+    moon = CIRS(cart, obstime=altaz.obstime)
+
+    moonaa = moon.transform_to(altaz)
+    assert np.abs(moonaa.distance - moon.distance).to(u.km) > 1000*u.km
+    assert np.abs(moonaa.distance - moon.distance).to(u.km) < 7000*u.km
